Refactor iDoc listener
