+++ native/src/main/java/io/ballerina/lib/sap/idochandlers/BallerinaIDocHandlerFactory.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright (c) 2024, WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.lib.sap.idochandlers;
+
+import com.sap.conn.idoc.jco.JCoIDocHandler;
+import com.sap.conn.idoc.jco.JCoIDocHandlerFactory;
+import com.sap.conn.idoc.jco.JCoIDocServerContext;
+import io.ballerina.runtime.api.Runtime;
+import io.ballerina.runtime.api.values.BObject;
+
+public class BallerinaIDocHandlerFactory implements JCoIDocHandlerFactory {
+
+    private final JCoIDocHandler handler;
+
+    public BallerinaIDocHandlerFactory(BObject service, Runtime runtime) {
+        this.handler = new BallerinaIDocHandler(service, runtime);
+    }
+
+    public JCoIDocHandler getIDocHandler(JCoIDocServerContext serverCtx) {
+        return handler;
+    }
+}
