+++ native/src/main/java/io/ballerina/lib/sap/idochandlers/BallerinaThrowableListener.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2024, WSO2 LLC. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.lib.sap.idochandlers;
+
+import com.sap.conn.jco.server.JCoServer;
+import com.sap.conn.jco.server.JCoServerContextInfo;
+import com.sap.conn.jco.server.JCoServerErrorListener;
+import com.sap.conn.jco.server.JCoServerExceptionListener;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class BallerinaThrowableListener implements JCoServerErrorListener, JCoServerExceptionListener {
+
+    private static final Logger logger = LoggerFactory.getLogger(BallerinaThrowableListener.class);
+
+    @Override
+    public void serverErrorOccurred(JCoServer jCoServer, String s, JCoServerContextInfo jCoServerContextInfo,
+                                    Error error) {
+        logger.error("Server error occurred: {}", error.getMessage());
+    }
+
+    @Override
+    public void serverExceptionOccurred(JCoServer jCoServer, String s, JCoServerContextInfo jCoServerContextInfo,
+                                        Exception e) {
+        logger.error("Server exception occurred: {}", e.getMessage());
+    }
+}
